+++ app/models/user.py
@@ -1,10 +1,7 @@
-from flask import Flask, render_template, redirect, url_for, request, session, flash
 from flask_sqlalchemy import SQLAlchemy
 
-app = Flask(__name__)
-app.secret_key = 'your_secret_key'
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///users.db'
-db = SQLAlchemy(app)
+db = SQLAlchemy()
+
 
 # User model
 class User(db.Model):
