+++ app/templates/base.html
@@ -1,5 +1,4 @@
-{% from 'macros/inline_delete.html' import delete_button %}
-{% from 'macros/inline_edit.html' import edit_button %}
+{% from 'macros/buttons.html' import add_button, edit_button, delete_button %}
 
 <!DOCTYPE html>
 <html lang="en">
@@ -16,14 +15,14 @@
         <div><h1>musico</h1></div>
         <nav>
             {% if 'user_id' in session %}
-                <a href="{{ url_for('home') }}">Home</a>
-                <a href="{{ url_for('clients') }}">Clients</a>
-                <a href="{{ url_for('profile') }}">Profile</a>
-                <a href="{{ url_for('logout') }}">Logout</a>
+                <a href="{{ url_for('users.home') }}">Home</a>
+                <a href="{{ url_for('clients.clients') }}">Clients</a>
+                <a href="{{ url_for('users.profile') }}">Profile</a>
+                <a href="{{ url_for('users.logout') }}">Logout</a>
             {% else %}
-                <a href="{{ url_for('pre_home') }}">Home</a>
-                <a href="{{ url_for('login') }}">Login</a>
-                <a href="{{ url_for('register') }}">Register</a>
+                <a href="{{ url_for('index.home') }}">Home</a>
+                <a href="{{ url_for('index.login') }}">Login</a>
+                <a href="{{ url_for('index.register') }}">Register</a>
             {% endif %}
         </nav>
     </div>
