+++ app/templates/clients/clients.html
@@ -9,8 +9,8 @@
     <h1>Clients</h1>
     <div class="page_buttons">
         
-        <a href="{{ url_for('manage_clients') }}" title="Add, Edit or Delete Clients">Manage Clients</a>
-        <a href="{{ url_for('view_clients') }}" title="View roster to avoid making unwanted changes">Roster</a>
+        <a href="{{ url_for('clients.manage_clients') }}" title="Add, Edit or Delete Clients">Manage Clients</a>
+        <a href="{{ url_for('clients.view_clients') }}" title="View roster to avoid making unwanted changes">Roster</a>
     </div>
    
     <div>
