+++ app/templates/clients/create_clients.html
@@ -6,8 +6,8 @@
 <main>
 
     <div class="page_buttons">
-        <a href="{{ url_for('manage_clients') }}">Manage Clients</a>
-        <a href="{{ url_for('view_clients') }}">Roster</a>
+        <a href="{{ url_for('clients.manage_clients') }}">Manage Clients</a>
+        <a href="{{ url_for('clients.view_clients') }}">Roster</a>
     </div>
     <h1>Create Client</h1>
     <form method="POST">
