+++ app/templates/clients/manage_clients.html
@@ -8,7 +8,7 @@
 
     <div class="page_buttons">
         <span class="tooltip_left" data-text="View Roster to avoid making unwanted changes.">
-            <a href="{{ url_for('view_clients') }}" class="page_buttons">Roster</a>
+            <a href="{{ url_for('clients.view_clients') }}" class="page_buttons">Roster</a>
         </span>
     </div>
 
@@ -16,7 +16,7 @@
     <caption>Client Roster</caption>
     <thead>
         <tr>
-            <th><a href="{{ url_for('create_client') }}"><img src="{{ url_for('static', filename='images/add.svg')}}" title="Add new client"/></a>
+            <th>{{ add_button(url_for('clients.create_client'))}}
             <th>Name</th>
             <th>Email</th>
             <th>Event Address</th>
@@ -38,10 +38,9 @@
                 <td>{{ client.date }}</td>
                 <td>
                     <!-- Button to edit the client -->
-                     {{ edit_button(client.name, url_for('update_client', client_id=client.id)) }}
-                    <!-- <a href="{{ url_for('update_client', client_id=client.id) }}"><img src="{{ url_for('static', filename='images/edit.svg')}}" title="Edit client details"/></a> -->
+                     {{ edit_button(client.name, url_for('clients.update_client', client_id=client.id)) }}
                     <!-- Button to delete the client -->
-                     {{ delete_button(client.name, url_for('delete_client', client_id=client.id)) }}
+                     {{ delete_button(client.name, url_for('clients.delete_client', client_id=client.id)) }}
                 </td>
             </tr>
         {% endfor %}
