+++ app/templates/clients/update_clients.html
@@ -5,7 +5,7 @@
 {% block content %}
 <main>
     <div class="page_buttons">
-        <a href="{{ url_for('clients') }}">Clients Home</a>
+        <a href="{{ url_for('clients.clients') }}">Clients Home</a>
     </div>
 <h1>Update Client</h1>
