+++ app/templates/clients/view_clients.html
@@ -6,7 +6,7 @@
 <main>
 <h1>Clients</h1>
     <div class="page_buttons">
-        <a href="{{ url_for('manage_clients') }}">Manage Clients</a>
+        <a href="{{ url_for('clients.manage_clients') }}">Manage Clients</a>
     </div>
 <div class="container">
     <div>
