--- app/templates/macros/inline_edit.html
@@ -1,4 +1,0 @@
-{% macro edit_button(entity_id, target_url) -%}
-<a href="{{ target_url }}" method="post" title="Edit entity details">
-    <img src="{{ url_for('static', filename='images/edit.svg')}}"/></a>
-{%- endmacro %}
