+++ app/templates/users/profile.html
@@ -6,7 +6,7 @@
 <main>
 <h1>{{user.username}}'s Profile</h1>
 <div class="page_buttons">
-    <a href="{{ url_for('update') }}">Update Profile</a>
+    <a href="{{ url_for('users.update') }}">Update Profile</a>
 </div>
 <table>
     <tr>
